Removed swiftData import and moved the DatabaseError enum
