+++ PVPCPlanner/App/PVPCPlannerApp.swift
@@ -1,5 +1,4 @@
 import SwiftUI
-import SwiftData
 
 @main
 struct PVPCPlannerApp: App {
