+++ PVPCPlanner/Data/DBContainers/PVPCDatabaseError.swift
@@ -1,0 +1,6 @@
+enum PVPCDatabaseError: Error {
+    case errorInsert
+    case errorFetch
+    case errorDelete
+    case errorUpdate
+}
